Update FriendPersistentInfo for 3DS
One of the DateTimes doesn't represent the date of the friendship, but
when the friend has updated their Mii.

I noticed this when playing with the friend servers researching
information about the data stored in the FRD sysmodule.

I will submit PRs to nex-protocols-go and the friends server to fix the
inaccuracy.