Add Real Escape Game docs
This adds documentation for the custom methods used in DataStore and
MatchmakeExtension on Real Esacape Game.

Alongside that, update the UserMessage structure to have the original
version and the NEX 4.0 version, and update the
DeliverMessageMultiTarget method on Messaging and Message Delivery with
field names from Real Escape Game.

The Xenoblade X page for Message Delivery is removed due to the method
being present in more games, and a minor fix on DataStore.