feat(matchmake-extension): Add documentation for Super Smash Bros 4
Contains the structures and method requests and responses that I was
able to locate inside the code by crossreferencing functions with the
retail version of Minecraft, and with help of the type_info tables.

Since the functions for extracting and adding signed or unsigned values
are indistinguishable from each other, it has been assumed for every
value to be unsigned. Keep in mind that this may not be the case, and
further research is needed!

There are various instances where it is unknown whether a List<u8> or a
Buffer should be used. In those cases a List<u8> is assumed, but this
may not be correct.

The fields details are mostly unknown. Only some of them have been
filled based on guesses from the method names. The PID is equivalent in
code to a Uint32, so there may also be instances of unknown fields which
are a Uint32 where a PID is used.