feat(subscriber): Improve documentation
All names are extracted from:
https://github.com/tech-ticks/RTDXTools/blob/232a7797e01369e9c12704f58fdde65dd3ac1c32/Assets/Scripts/Stubs/Generated/Assembly-CSharp/NexPlugin/Subscriber.cs

Some of the types look suspicious, like the "topics" being strings
instead of numbers like on the C# code, or the UpdateUserStatus method,
which doesn't seem to match. I assume this may be because C# is
abstracting those differences, so the types are kept the same, and the
names are put on top of the fields instead.