feat(nex): Add Rating protocol
This is the protocol used in Pokemon games. The types and structure
names were extracted from the game binaries, thanks to Pokemon games
having type_info data.

The code shows that this protocol has 9 methods. However, there is only
code for calling 7 of them (6 and 9 are missing), as they are unused by
the games.

The List<Uint8> types may not be accurate, as they are indistinguishable
in code from Buffers. I have chosen to document them as lists to be
safe, but it should be kept in mind when implementing the protocol.

The protocol has evolved over time. For example, Pokemon X only has the
first 3 methods, where as Pokemon Ultra Sun contains all of of the
methods documented here.