feat: Expand legacy Ranking and Storage Manager documentation
The new information about the methods and their parameters has been
obtained by reverse-engineering games which use legacy Ranking. This
helps determine the real arguments of various methods, since these
protocols didn't validate at the time that all the parameters were being
provided properly.

Alongside that, this allows figuring out more information about the
response parameters on some methods for which they were unknown.
Interestingly, on method 0x13 of Ranking, the game only extracts the
"result code", but this was documented at the time with an unknown
format. It is unknown what those other parameters not being accounted
here are, and the result code for this method has unfortunatelly been
lost.

The new names assigned to the methods and structures are contextual
guesses, and may not represent their real names. Similarly, like in
other reverse-engineering efforts, the new parameters may be either
signed or unsigned, but this is unknown given the assembly treats them
the same way, so we are assuming all of them will be unsigned.

As an additional note, the methods documented as `GetScore` and
`GetSelfScore` get called from the same function internally, and which
one is called is determined by the ranking mode (if set to 1, it will
call `GetSelfScore`, otherwise it calls `GetScore`).