feat(kerberos): Expand NEX 4 documentation
Correct some documentation which was distinguished as a platform
difference when in fact it's a NEX version difference. Also add new
undocumented ticket format only seen in NEX 4 games.

Finally, add documentation regarding NEX functionality for crossplay
games. This has been extracted from Monster Hunter Generations Ultimate
and its data transfer tool, but given the generic naming of the fields
being added I'm confident this is a feature in NEX aswell and not
specific to this game, even if we don't have any other examples about
it.